# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Orophile Trek API",
        "description": "API documentation for Orophile Trek application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/treks": {
            "get": {
                "tags": [
                    "Treks"
                ],
                "summary": "Get list of treks",
                "operationId": "0035c0c9cbf3f20405ebffa61c2d85ae",
                "parameters": [
                    {
                        "name": "data_type",
                        "in": "query",
                        "description": "Filter treks by type (trek or package)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "trek",
                                "package"
                            ]
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filter by active status",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Treks"
                ],
                "summary": "Create a new trek",
                "operationId": "5f2574c5537571e2e9ae59b5d4184dc3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "location",
                                    "price",
                                    "currency",
                                    "duration",
                                    "difficulty",
                                    "type",
                                    "distance_km",
                                    "data_type",
                                    "is_featured",
                                    "is_active",
                                    "trek_days"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "data_type": {
                                        "type": "string",
                                        "enum": [
                                            "trek",
                                            "package"
                                        ]
                                    },
                                    "location": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0
                                    },
                                    "currency": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "duration": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "difficulty": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "type": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "distance_km": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "featured_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "is_featured": {
                                        "type": "boolean"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "trek_days": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Trek created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Trek created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/treks/{id}": {
            "get": {
                "tags": [
                    "Treks"
                ],
                "summary": "Get a single trek",
                "operationId": "defad44ab2d7dadc9df1e63182d739ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Treks"
                ],
                "summary": "Update a trek",
                "operationId": "c01e60d5fa1ed074a8ec211e12f18c6c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "data_type": {
                                        "type": "string",
                                        "enum": [
                                            "trek",
                                            "package"
                                        ]
                                    },
                                    "location": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0
                                    },
                                    "currency": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "duration": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "difficulty": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "type": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "distance_km": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "featured_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "is_featured": {
                                        "type": "boolean"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "trek_days": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Trek updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Treks"
                ],
                "summary": "Delete a trek",
                "operationId": "4b2e5d32d06f75317c966cce9be98215",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}